refactor: replace PNG logos with WEBP format for Kimi and Moonshot
- Deleted outdated PNG files for Kimi, Moonshot, and their respective models.
- Added new WEBP format logos for Kimi and Moonshot in the assets.
- Updated import paths in configuration files to reference the new WEBP logos.
